2010-02-25  Maurits Rijk  <maurits.rijk@gmail.com>
	* lib/FuzzySelectTool.cs: bug fix

	* plug-ins/PhotoshopActions/SelectPointEvent.cs: minor clean-up

	* plug-ins/PhotoshopActions/Abbreviations.cs: added abbreviation

+++ lib/Pixel.cs
@@ -137,7 +137,7 @@ namespace Gimp
     {
       get 
 	{
-	  return new RGB(Red, Green, Blue);
+	  return new RGB((byte) Red, (byte) Green, (byte) Blue);
 	}
 
       set
